feat: use Into/TryInto traits to convert products to/from DynamoDB
