+++ src/store/dynamodb/mod.rs
@@ -6,11 +6,11 @@ use super::{Store, StoreDelete, StoreGet, StoreGetAll, StorePut};
 use crate::{Error, Product, ProductRange};
 use async_trait::async_trait;
 use aws_sdk_dynamodb::{model::AttributeValue, Client};
-use std::str;
+use std::collections::HashMap;
 use tracing::{info, instrument};
 
 mod ext;
-use ext::{AttributeValuesExt, ProductsExt};
+use ext::{AttributeValuesExt};
 
 /// DynamoDB store implementation.
 ///
@@ -55,7 +55,7 @@ where
         let products = match res.items {
             Some(items) => items
                 .into_iter()
-                .map(Product::from_dynamodb)
+                .map(|v| v.try_into())
                 .collect::<Result<Vec<Product>, Error>>()?,
             None => Vec::default(),
         };
@@ -82,7 +82,7 @@ where
             .await?;
 
         Ok(match res.item {
-            Some(item) => Some(Product::from_dynamodb(item)?),
+            Some(item) => Some(item.try_into()?),
             None => None,
         })
     }
@@ -100,7 +100,7 @@ where
         self.client
             .put_item()
             .table_name(&self.table_name)
-            .set_item(Some(product.to_dynamodb()))
+            .set_item(Some(product.into()))
             .send()
             .await?;
 
@@ -125,6 +125,42 @@ where
             .await?;
 
         Ok(())
+    }
+}
+
+impl Into<HashMap<String, AttributeValue>> for &Product {
+    /// Convert a &Product into a DynamoDB item
+    fn into(self) -> HashMap<String, AttributeValue> {
+        let mut retval = HashMap::new();
+        retval.insert("id".to_owned(), AttributeValue::S(self.id.clone()));
+        retval.insert("name".to_owned(), AttributeValue::S(self.name.clone()));
+        retval.insert(
+            "price".to_owned(),
+            AttributeValue::N(format!("{:}", self.price)),
+        );
+
+        retval
+    }
+}
+
+impl TryInto<Product> for HashMap<String, AttributeValue> {
+    type Error = Error;
+
+    /// Try to convert a DynamoDB item into a Product
+    /// 
+    /// This could fail as the DynamoDB item might be missing some fields.
+    fn try_into(self) -> Result<Product, Self::Error> {
+        Ok(Product {
+            id: self
+                .get_s("id")
+                .ok_or(Error::InternalError("Missing id"))?,
+            name: self
+                .get_s("name")
+                .ok_or(Error::InternalError("Missing name"))?,
+            price: self
+                .get_n("price")
+                .ok_or(Error::InternalError("Missing price"))?,
+        })
     }
 }
 
@@ -329,6 +365,35 @@ mod tests {
 
         // THEN the request matches the expected request
         conn.assert_requests_match(&vec![]);
+
+        Ok(())
+    }
+
+    #[test]
+    fn product_from_dynamodb() {
+        let mut value = HashMap::new();
+        value.insert("id".to_owned(), AttributeValue::S("id".to_owned()));
+        value.insert("name".to_owned(), AttributeValue::S("name".to_owned()));
+        value.insert("price".to_owned(), AttributeValue::N("1.0".to_owned()));
+
+        let product: Product = value.try_into().unwrap();
+        assert_eq!(product.id, "id");
+        assert_eq!(product.name, "name");
+        assert_eq!(product.price, 1.0);
+    }
+
+    #[test]
+    fn product_to_dynamodb() -> Result<(), Error> {
+        let product = Product {
+            id: "id".to_owned(),
+            name: "name".to_owned(),
+            price: 1.5,
+        };
+
+        let value: HashMap<String, AttributeValue> = (&product).into();
+        assert_eq!(value.get("id").unwrap().as_s().unwrap(), "id");
+        assert_eq!(value.get("name").unwrap().as_s().unwrap(), "name");
+        assert_eq!(value.get("price").unwrap().as_n().unwrap(), "1.5");
 
         Ok(())
     }
